requirements.txt
Brotli==1.1.0
defusedxml==0.7.1
fonttools==4.56.0
fpdf==1.7.2
fpdf2==2.8.2
inflate64==1.0.1
lxml==5.3.1
multivolumefile==0.2.3
numpy==2.2.2
pandas==2.2.3
pillow==11.1.0
psutil==6.1.1
py7zr==0.22.0
pybcj==1.0.3
pycryptodome==3.21.0
pycryptodomex==3.21.0
pyppmd==1.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
pyzstd==0.16.2
six==1.17.0
texttable==1.7.0
typing_extensions==4.12.2
tzdata==2025.1
